pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semindex"
version = "0.1.0"
description = "Local semantic codebase indexer (Python-only, CPU-only)"
authors = [{ name = "You" }]
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "faiss-cpu==1.8.0",
    "transformers==4.44.2",
    "torch==2.4.0",
    "sentencepiece==0.2.0",
    "tqdm==4.66.5",
    "numpy==1.26.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
semindex = "semindex.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
